[genkey] Add search for hexadecimal English words
Added a command-line option `-words <#>' that allows genkey to search
for generated keys that resolve to IP addresses with four-character
hexadecimal "words." The numeric value in the option specified the
number of words minimum to find in the first half of the IP address. For
example, searching for IP addresses with one word generates the
following output with the first IP address containing the word "feed"
and the second address containing the word "bead":

    # go run ./cmd/genkeys -words 1
    Threads: 12 Minimum Words: 1
    ----- 14.304191ms
    Priv: 4d43...
    Pub: 40449002ea06e94020117c260c3bd2bcde1d21c41b2819e2f4a2090212fc3902
    IP: 201:feed:bff4:57e4:5aff:7fba:f67:cf10
    ----- 74.741243ms
    Priv: 137b...
    Pub: 288c632f882a5afa5f7515f00a2e18328473709b2b69ba3540602345141cb54c
    IP: 202:bb9c:e683:bead:282d:457:507f:ae8f

The default value if the `-words` option is not use is zero, meaning no
words are necessary to successfully match an address. This maintains the
original function of the program. Words are only searched for in the
first half of the IP address as this is the part of the address that
remains permanent for the local IP address.

The wordlist consists of:
    * bead
    * beef
    * dead
    * deaf
    * deed
    * face
    * feed

This is not intended to be a comprehensive list, and probably could be
written better in an expandible array. It serves it's purpose for now.